requirements.txt
discord.py>=2.3.2
python-dotenv>=0.21.0
